Extend code generation to smallvec, fix append, optimize extend, bug in reserve+set
Fixed append because smallvec does not have that method, and I forgot to clear the other vector, now there is a test for that. extend() was optimized to reserve+set_raw instead of reserve+push, but it turns out that reserve does not change the size of the storage, so when doing a clone it will not clone anything because the storage has len = 0. A possible fix would be to fix storage before doing the clone, but the fix here is to resize the storage on reserve(). There is a test for that now.